Add mechanism to reset an account before e2e tests
Adds a mechanism to remove all keys/devices of a specified account in
order to start all tests with a known state. This is achieved using a
simple http client included in the e2e test apk, which support login,
listing devices and removing devices.

An account used for testing _must_ be specified either as a gradle
property or as a runtime to the test runner:
* Local property (local.properties):
  test_account_token=XXXXXXXXXXXXXXXX
* Command line argument:
  ./gradlew :e2e:assembleDebug -Ptest_account_token=XXXXXXXXXXXXXXXX
* Runtime argument:
  am instrument -e test_account_token XXXXXXXXXXXXXXXX # ...